further tidying
